Add files via upload
In this project I use Random Forest Classification as well as KNN, and K-fold Classifier in order to compare the score and accuracy of each model thus being aware of overfitting. 
I used the local Internet Server Data Set in order to train and test my models.  
At the end the goal is to predict a WIFI user's locations based on the wifi signal strength recorder from the local ISP. 
The accuracy achieved is around 97%.